feat(bash): support path input for `dus` alias
The `dus` alias, defined in `/.arch_aliases`, is a wrapper around `du`
and `sort`. As an alias, trying to provide a path as input for `du`
would instead pass it to `sort`. This changes `dus` to a function that
passes through the first argument it receives to `du` so it can be used
to scan directories other than the current working directory.